Cambio de float a int
Se hizo un cambio de float a int así disminuimos el tamaño del
programa.

+++ conversionCF.c
@@ -7,11 +7,11 @@ int main() {
   //creamos 4 variables f,divi,multi,resta
   //f es la variable que se va a ingresar
   //Todas las variables estan como float
-  float f, divi=0,multi = 0,resta = 0;
+  int f, divi=0,multi = 0,resta = 0;
 
 
   printf("Ingrese los grados F: ");
-  scanf("%f", &f); //pedimos el valor al usuario
+  scanf("%i", &f); //pedimos el valor al usuario
 
   //ecuaciones para calcular a grados C
   resta = f - 32; //valor f resta 32
@@ -19,7 +19,7 @@ int main() {
   divi = multi / 9; //multi se divide entre 9
   //restultado se muestra abajo
   //reducimos el float a 2 digitos despues de la coma
-  printf("Son en grados C: %.2f\n", divi);
+  printf("Son en grados C: %i\n", divi);
 
   return 0;
 }
